enforced float type for csu bringi kdp

+++ openradartools/dp.py
@@ -226,8 +226,8 @@ def phidp_bringi(radar, gatefilter, phidp_field="PHI_UNF", refl_field='DBZ'):
     kdpb: ndarray
         Bringi specific differential phase array.
     """
-    dz = radar.fields[refl_field]['data'].copy().filled(-9999)
-    dp = radar.fields[phidp_field]['data'].copy().filled(-9999)
+    dz = radar.fields[refl_field]['data'].copy().filled(-9999).astype(np.single)
+    dp = radar.fields[phidp_field]['data'].copy().filled(-9999).astype(np.single)
 
     # Extract dimensions
     rng = radar.range['data']
@@ -236,7 +236,7 @@ def phidp_bringi(radar, gatefilter, phidp_field="PHI_UNF", refl_field='DBZ'):
     [R, A] = np.meshgrid(rng, azi)
 
     # Compute KDP bringi.
-    kdpb, phidpb, _ = csu_kdp.calc_kdp_bringi(dp, dz, R / 1e3, gs=dgate, bad=-9999, thsd=12, window=6.0, std_gate=11)
+    kdpb, phidpb, _ = csu_kdp.calc_kdp_bringi(dp, dz, R.astype(np.single) / 1e3, gs=float(dgate), bad=-9999, thsd=12.0, window=6.0, std_gate=11)
 
     # Mask array
     phidpb = np.ma.masked_where(phidpb == -9999, phidpb)
